requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.3
cssselect==1.2.0
cssutils==2.9.0
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
emails==0.6
fastapi==0.109.2
fastapi-cli==0.0.3
filelock==3.13.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.25.2
identify==2.5.35
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
lxml==5.1.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.10.3
packaging==24.0
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
premailer==3.10.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
pyasn1==0.6.0
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pytest==7.4.4
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
ruff==0.2.2
sentry-sdk==1.41.0
setuptools==69.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
tenacity==8.2.3
typer==0.12.3
types-passlib==1.7.7.20240311
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.24.0.post1
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
